Q4_THINKING/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.20.0